Adds hierarchical sheets, picks PC0 for ADC
Adds CAN & power distribution sheets. The VoltageDivision sheet is probably not needed (it uses the TPS3711 IC as a voltage comparator). PC0 is picked to be the ADC input (but we probably won't use the voltage divider as it is on the schematic right now, an op amp will replace the voltage divider).